Added AudioLayer sampler management routines into ClarkMix
git-svn-id: file:///home/chrism/dev/old-cvs/old-cvs/trunk/looper@113 eec3997f-6950-4fa9-abeb-60fabf57e09c

+++ Sample.cc
@@ -93,7 +93,7 @@ char *Sample::GetName()
 }
 
 // set the size of the array of data to be returned by GetChunk
-void Sample::SetChunkSize(u16 setsize)
+void Sample::SetChunkSize(u8 setsize)
 {
 	chunksize = setsize;
 }
